feat: Splitted registerController from UsersController
