+++ src/application/controllers/__tests__/UsersController.test.ts
@@ -1,11 +1,8 @@
 import { mocked } from 'ts-jest/utils'
 import User from '@domain/entities/User'
-import CreateUserRequest from '@application/resources/CreateUserRequest'
-import register from '@domain/use_cases/register'
 import {
   deleteUserController,
   loginController,
-  registerController,
   updateUserController
 } from '@application/controllers/UsersController'
 import LoginRequest from '@application/resources/LoginRequest'
@@ -22,45 +19,15 @@ import UpdateUserRequest from '@application/resources/UpdateUserRequest'
 import UpdateUserResponse from '@application/resources/UpdateUserResponse'
 import updateUser from '@domain/use_cases/updateUser'
 
-jest.mock('@domain/use_cases/register')
 jest.mock('@domain/use_cases/login')
 jest.mock('@domain/use_cases/deleteUser')
 jest.mock('@domain/use_cases/updateUser')
 
-const mockedRegister = mocked(register)
 const mockedLogin = mocked(login)
 const mockedDeleteUser = mocked(deleteUser)
 const mockedUpdateUser = mocked(updateUser)
 
 describe('Controllers: Users Controller', () => {
-  describe('Register Controller', () => {
-    describe('when sending a valid request', () => {
-      it('calls register use case with all params', async () => {
-        const request = new CreateUserRequest({
-          username: 'test',
-          email: 'test@test.com',
-          password: '123456'
-        })
-        const user = new User(
-          '613c1a2a7c3c65de8fa77384',
-          request.username,
-          request.email,
-          request.password
-        )
-        mockedRegister
-          .mockImplementation(async () => user)
-
-        await registerController(request)
-
-        expect(register).toHaveBeenCalledWith(
-          request.username,
-          request.email,
-          request.password
-        )
-      })
-    })
-  })
-
   describe('Login Controller', () => {
     describe('when sending a valid request', () => {
       it('calls login use case with all params', async () => {
